添加计算器UI

+++ No_rz43_CalcualtorLib/Calculators/Add.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 
 namespace No_rz43_CalcualtorLib.Calculators
 {
-    public class Add: ICalculator
+    public class Add : ICalculator
     {
-        public int Calcualtor(int var1, int var2)
+        public double Calcualtor(double var1, double var2)
         {
             var result = var1 + var2;
             return result;
