+++ No_rz43_CalcualtorLib/Calculators/Div.cs
@@ -8,16 +8,15 @@ namespace No_rz43_CalcualtorLib.Calculators
 {
     public class Div : ICalculator
     {
-        public int Calcualtor(int var1, int var2)
+        public double Calcualtor(double var1, double var2)
         {
             if (var2 != 0)
             {
-                int result = var1  / var2;
+                double result = var1 / var2;
                 return result;
             }
             else
                 throw new Exception("除数不能为零...");
         }
-
     }
 }
