+++ No_rz43_CalcualtorLib/Calculators/Mul.cs
@@ -8,11 +8,10 @@ namespace No_rz43_CalcualtorLib.Calculators
 {
     public class Mul : ICalculator
     {
-        public int Calcualtor(int var1, int var2)
+        public double Calcualtor(double var1, double var2)
         {
             var result = var1 * var2;
             return result;
         }
-
     }
 }
