+++ No_rz43_CalcualtorLib/Calculators/Sub.cs
@@ -8,7 +8,7 @@ namespace No_rz43_CalcualtorLib.Calculators
 {
     public class Sub : ICalculator
     {
-        public int Calcualtor(int var1, int var2)
+        public double Calcualtor(double var1, double var2)
         {
             var result = var1 - var2;
             return result;
